+++ client/src/app/cms/pages/edit-customer/edit-customer.component.ts
@@ -1,13 +1,11 @@
 import { Component, OnInit } from '@angular/core';
-import { Router, ActivatedRoute, ActivationEnd } from '@angular/router';
+import { Router, ActivatedRoute } from '@angular/router';
 import { switchMap } from 'rxjs/operators';
 import { NgForm } from '@angular/forms';
 import { CustomerService } from 'src/app/services/customer.service';
-import { UserService } from 'src/app/services/user.service';
+import { AlertsService } from 'src/app/services/alerts.service';
 import {
   Customer,
-  CreateCustomerDTOByAdmin,
-  CreateCustomerDTO,
   UpdateCustomerDTO,
 } from '../../../models/customer.model';
 
@@ -17,7 +15,6 @@ import {
   styleUrls: ['./edit-customer.component.css'],
 })
 export class EditCustomerComponent implements OnInit {
-  idUsuario: number = 0;
   client: Customer = {
     id: 0,
     name: '',
@@ -36,7 +33,7 @@ export class EditCustomerComponent implements OnInit {
     private customerService: CustomerService,
     private router: Router,
     private route: ActivatedRoute,
-    private userService: UserService
+    private alertsService: AlertsService,
   ) {
     this.isDisabled = true;
   }
@@ -66,6 +63,7 @@ export class EditCustomerComponent implements OnInit {
   }
   editarCliente(f: NgForm) {
     if (!f.valid) {
+      this.alertsService.alertaFailTop('top-end','error','Error!!','Formulario no válido',false,1500);
     } else {
       const dto: UpdateCustomerDTO = {
         name: this.client.name,
@@ -78,18 +76,26 @@ export class EditCustomerComponent implements OnInit {
       };
 
       this.customerService
-        .updateClient(this.client.id, dto).subscribe((client) => {
-          console.log('cliente actualizado', client);
+        .updateClient(this.client.id, dto).subscribe(() => {
+          this.alertsService.alertaSuccessTop('top-end','success','Cliente modificado',false,1500);
           this.router.navigate(['/cms/customers']);
-        });
+        },(()=>{
+          this.alertsService.alertaFailTop('top-end','error','Error!!','Error al editar cliente',false,1500);
+        }));
     }
   }
 
   deleteClient() {
-    this.customerService.deleteClient(this.client.id).subscribe(data=>{
-      console.log('client eliminado:', data);
-      this.router.navigate(['/cms/customers']);
+    this.alertsService.alertaDelete('Estas seguro?','No podrás revertir los cambios','warning',true,'#3085d6',
+    '#d33','Si, eliminar').then((result) => {
+      if (result.isConfirmed) {
+        this.customerService.deleteClient(this.client.id).subscribe(data=>{
+          this.alertsService.alertaSuccessTop('top-end','success','Cliente eliminado',false,1500);
+          this.router.navigate(['/cms/customers']);
+        });
+      }
     });
+    
   }
   habilitar() {
     this.isDisabled = false;
